+++ src/App.js
@@ -1,35 +1,21 @@
-import './components/category-item/category-item.component.jsx'
-import Directory from './components/directory/directory.component';
+import "./components/category-item/category-item.component.jsx";
+import Home from "./routes/home/home.component.jsx";
+import { Routes, Route} from "react-router-dom";
+import Navigation from "./routes/navigation/navigation.component.jsx";
+
+const Shop = () => {
+  return <h1>Shop</h1>;
+};
+
+
 const App = () => {
-  const categories = [
-    {
-      id: 1,
-      title: "hats",
-      imageUrl: "https://i.ibb.co/cvpntL1/hats.png",
-    },
-    {
-      id: 2,
-      title: "jackets",
-      imageUrl: "https://i.ibb.co/px2tCc3/jackets.png",
-    },
-    {
-      id: 3,
-      title: "sneakers",
-      imageUrl: "https://i.ibb.co/0jqHpnp/sneakers.png",
-    },
-    {
-      id: 4,
-      title: "women's",
-      imageUrl: "https://i.ibb.co/GCCdy8t/womens.png",
-    },
-    {
-      id: 5,
-      title: "men's",
-      imageUrl: "https://i.ibb.co/R70vBrQ/men.png",
-    },
-  ];
   return (
-    <Directory categories={categories}/>
+    <Routes>
+      <Route path="/" element={<Navigation/>}>
+        <Route index element={<Home />}/>
+        <Route path="shop" element={<Shop />} />
+      </Route>
+    </Routes>
   );
 };
