+++ src/components/category-item/category-item.component.jsx
@@ -1,4 +1,4 @@
-import './category-item.styles.scss'
+import "./category-item.styles.scss";
 
 const CategoryItem = ({category}) =>{
     const {imageUrl, title} = category;
@@ -15,7 +15,7 @@ const CategoryItem = ({category}) =>{
           <p>Shop Now</p>
         </div>
       </div>
-    )
+    );
 }
 
-export default CategoryItem
+export default CategoryItem;
