+++ src/index.scss
@@ -10,3 +10,12 @@ code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
 }
+
+a{
+  text-decoration:none;
+  color : black
+}
+
+* {
+  box-sizing: border-box;
+}
